requirements.txt
amqp==2.5.2
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==2.1.1
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
wrapt==1.11.2
zipp==3.1.0
